Prototype control plot code
This update changes the way the datacard model works, adding an option for control plot mode, and changing where it looks for shapes, and how a few shapes work. It also changes the running script to have a --ControlMode option for running this sort of action. A simple drawing script is also included using some work from the prefit modules.